Extract/Add some more concepts to make updates and user interactions easier to deal with.
Details:
add the concept of an input observer. Classes that implement this
interface are expected to watch for controller input on each frame.

add an InteractionManager to so there's one place to deal with click,
hover, etc detection.

add an Entity base class. All objects that can be manipulated in the game
will eventually inherit from this.

add an entity list to the game class. This list contains all game entities
(things that can be clicked, drawn, etc). By putting all the entities into
one list we can easily perform operations on all of them (or subsets of
them) at once.

Store button now implements IInputObservable and delegates to
InteractionManager for click detection.